+++ embiggen/__init__.py
@@ -1,8 +1,8 @@
-from .embedders import CBOW, SkipGram, GloVe
+from .embedders import CBOW, SkipGram, GloVe, BinarySkipGram
 from .transformers import NodeTransformer, EdgeTransformer, GraphTransformer
-from .sequences import NodeCBOWSequence, NodeSkipGramSequence, LinkPredictionSequence
+from .sequences import NodeBinarySkipGramSequence, Node2VecSequence, LinkPredictionSequence
 
 __all__ = [
-    "CBOW", "SkipGram", "GloVe", "NodeCBOWSequence",
-    "NodeSkipGramSequence", "LinkPredictionSequence"
+    "CBOW", "SkipGram", "GloVe", "BinarySkipGram",
+    "NodeBinarySkipGramSequence", "LinkPredictionSequence", "Node2VecSequence"
 ]
