+++ embiggen/embedders/__init__.py
@@ -1,7 +1,8 @@
 from .glove import GloVe
 from .skipgram import SkipGram
+from .binary_skipgram import BinarySkipGram
 from .cbow import CBOW
 
 __all__ = [
-    "GloVe", "SkipGram", "CBOW"
+    "GloVe", "SkipGram", "CBOW", "BinarySkipGram"
 ]
