+++ embiggen/sequences/__init__.py
@@ -1,7 +1,7 @@
-from .node_cbow_sequence import NodeCBOWSequence
-from .node_skipgram_sequence import NodeSkipGramSequence
+from .node2vec_sequence import Node2VecSequence
+from .node_binary_skipgram_sequence import NodeBinarySkipGramSequence
 from .link_prediction_sequence import LinkPredictionSequence
 
 __all__ = [
-    "NodeCBOWSequence", "NodeSkipGramSequence", "LinkPredictionSequence"
+    "Node2VecSequence", "NodeBinarySkipGramSequence", "LinkPredictionSequence"
 ]
