+++ tests/test_abstract_node2vec_sequence.py
@@ -1,0 +1,21 @@
+import pytest
+from embiggen.sequences.abstract_node2vec_sequence import AbstractNode2VecSequence
+from .test_sequences import TestSequences
+
+
+class TestAbstractNode2VecSequence(TestSequences):
+
+    def setUp(self):
+        super().setUp()
+        self._window_size = 4
+        self._length = 100
+        self._batch_size = 1
+        self._sequence = AbstractNode2VecSequence(
+            self._graph,
+            length=self._length,
+            batch_size=self._batch_size,
+            window_size=self._window_size
+        )
+
+        with pytest.raises(NotImplementedError):
+            self._sequence.__getitem__(0)
